spring boot project day 3

+++ src/main/java/com/project/football_club_matching/entity/Example.java
@@ -28,7 +28,7 @@ public class Example {
     @JoinColumn(name = "QUESTION_ID")
     private Question question;
 
-    private Long score;
+    private int score;
 
     
     private String teams;
